Updated db call in search function
Now uses safe PreparedStatements, only searches for fields that have
text entered, and doesn't search if nothing is entered.
Probably should add a label to tell user if they try to search with no
parameters, a button to cancel updating a patient (resets patient info
fields and button disabled states), and possibly a button to show all
patients (if needed).